refactor: Phase 1 production hardening
- Migrate all monetary fields from float to BigDecimal
- Add @Transactional to DatabaseConduit.process()
- Add SLF4J structured logging across all components
- Add GlobalExceptionHandler with @RestControllerAdvice
- Add LocalDateTime timestamp to TransactionRecord
- Fix controller package structure

+++ src/test/java/com/jpmc/midascore/UserPopulator.java
@@ -1,5 +1,5 @@
 package com.jpmc.midascore;
-
+import java.math.BigDecimal;
 import com.jpmc.midascore.component.DatabaseConduit;
 import com.jpmc.midascore.entity.UserRecord;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +17,7 @@ public class UserPopulator {
         String[] userLines = fileLoader.loadStrings("/test_data/lkjhgfdsa.hjkl");
         for (String userLine : userLines) {
             String[] userData = userLine.split(", ");
-            UserRecord user = new UserRecord(userData[0], Float.parseFloat(userData[1]));
+        UserRecord user = new UserRecord(userData[0], BigDecimal.valueOf(Float.parseFloat(userData[1])));
             databaseConduit.save(user);
         }
     }
